+++ packages/client-core/src/systems/LoadingUISystem.tsx
@@ -58,8 +58,8 @@ import { CameraComponent } from '@etherealengine/engine/src/camera/components/Ca
 import { UUIDComponent } from '@etherealengine/engine/src/common/UUIDComponent'
 import { InputComponent } from '@etherealengine/engine/src/input/components/InputComponent'
 import { addObjectToGroup, GroupComponent } from '@etherealengine/engine/src/scene/components/GroupComponent'
+import { setObjectLayers } from '@etherealengine/engine/src/scene/components/ObjectLayerComponent'
 import { SceneSettingsComponent } from '@etherealengine/engine/src/scene/components/SceneSettingsComponent'
-import { setObjectLayers } from '@etherealengine/engine/src/scene/functions/setObjectLayers'
 import { TransformComponent } from '@etherealengine/engine/src/transform/components/TransformComponent'
 import { TransformSystem } from '@etherealengine/engine/src/transform/systems/TransformSystem'
 import { AdminClientSettingsState } from '../admin/services/Setting/ClientSettingService'
