+++ packages/editor/src/classes/TransformGizmoComponent.ts
@@ -38,10 +38,10 @@ import { CameraComponent } from '@etherealengine/engine/src/camera/components/Ca
 import { NameComponent } from '@etherealengine/engine/src/common/NameComponent'
 import { EngineRenderer } from '@etherealengine/engine/src/renderer/WebGLRendererSystem'
 import { addObjectToGroup, removeObjectFromGroup } from '@etherealengine/engine/src/scene/components/GroupComponent'
+import { setObjectLayers } from '@etherealengine/engine/src/scene/components/ObjectLayerComponent'
 import { VisibleComponent } from '@etherealengine/engine/src/scene/components/VisibleComponent'
 import { ObjectLayers } from '@etherealengine/engine/src/scene/constants/ObjectLayers'
 import { SnapMode, TransformPivot } from '@etherealengine/engine/src/scene/constants/transformConstants'
-import { setObjectLayers } from '@etherealengine/engine/src/scene/functions/setObjectLayers'
 import { TransformComponent } from '@etherealengine/engine/src/transform/components/TransformComponent'
 import { getMutableState, getState, useHookstate } from '@etherealengine/hyperflux'
 import { useEffect } from 'react'
