+++ packages/editor/src/components/properties/PostProcessingSettingsEditor.tsx
@@ -29,8 +29,8 @@ import { useTranslation } from 'react-i18next'
 import { Color } from 'three'
 
 import { useComponent } from '@etherealengine/ecs/src/ComponentFunctions'
+import { Effects } from '@etherealengine/engine/src/renderer/effects/PostProcessing'
 import { PostProcessingComponent } from '@etherealengine/engine/src/scene/components/PostProcessingComponent'
-import { Effects } from '@etherealengine/engine/src/scene/constants/PostProcessing'
 import AutoFixHighIcon from '@mui/icons-material/AutoFixHigh'
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp'
