+++ packages/engine/src/avatar/components/AvatarIKComponents.ts
@@ -35,9 +35,9 @@ import { NameComponent } from '../../common/NameComponent'
 import { NetworkObjectComponent } from '../../networking/components/NetworkObjectComponent'
 import { RendererState } from '../../renderer/RendererState'
 import { addObjectToGroup, removeObjectFromGroup } from '../../scene/components/GroupComponent'
+import { setObjectLayers } from '../../scene/components/ObjectLayerComponent'
 import { VisibleComponent } from '../../scene/components/VisibleComponent'
 import { ObjectLayers } from '../../scene/constants/ObjectLayers'
-import { setObjectLayers } from '../../scene/functions/setObjectLayers'
 import { TransformComponent } from '../../transform/components/TransformComponent'
 import { AvatarRigComponent } from './AvatarAnimationComponent'
