+++ packages/engine/src/avatar/functions/avatarFunctions.ts
@@ -38,8 +38,8 @@ import {
 } from '@etherealengine/ecs/src/ComponentFunctions'
 import { Entity } from '@etherealengine/ecs/src/Entity'
 import { AssetLoader } from '../../assets/classes/AssetLoader'
+import { setObjectLayers } from '../../scene/components/ObjectLayerComponent'
 import { ObjectLayers } from '../../scene/constants/ObjectLayers'
-import { setObjectLayers } from '../../scene/functions/setObjectLayers'
 import { computeTransformMatrix } from '../../transform/systems/TransformSystem'
 import { AnimationState } from '../AnimationManager'
 // import { retargetSkeleton, syncModelSkeletons } from '../animation/retargetSkeleton'
