+++ packages/engine/src/initializeEngine.ts
@@ -37,9 +37,9 @@ import { Timer } from '@etherealengine/ecs/src/Timer'
 import { getMutableState } from '@etherealengine/hyperflux'
 import { NameComponent } from './common/NameComponent'
 import { addObjectToGroup } from './scene/components/GroupComponent'
+import { setObjectLayers } from './scene/components/ObjectLayerComponent'
 import { VisibleComponent } from './scene/components/VisibleComponent'
 import { ObjectLayers } from './scene/constants/ObjectLayers'
-import { setObjectLayers } from './scene/functions/setObjectLayers'
 import { TransformComponent } from './transform/components/TransformComponent'
 import { XRState } from './xr/XRState'
