+++ packages/engine/src/mocap/solveMotionCapturePose.ts
@@ -63,9 +63,9 @@ import { NameComponent } from '../common/NameComponent'
 import { V_010, V_100 } from '../common/constants/MathConstants'
 import { RendererState } from '../renderer/RendererState'
 import { GroupComponent, addObjectToGroup } from '../scene/components/GroupComponent'
+import { setObjectLayers } from '../scene/components/ObjectLayerComponent'
 import { setVisibleComponent } from '../scene/components/VisibleComponent'
 import { ObjectLayers } from '../scene/constants/ObjectLayers'
-import { setObjectLayers } from '../scene/functions/setObjectLayers'
 import { TransformComponent } from '../transform/components/TransformComponent'
 import { MotionCaptureRigComponent } from './MotionCaptureRigComponent'
