+++ packages/engine/src/renderer/WebGLRendererSystem.ts
@@ -51,11 +51,11 @@ import { CameraComponent } from '../camera/components/CameraComponent'
 import { ExponentialMovingAverage } from '../common/classes/ExponentialAverageCurve'
 import { overrideOnBeforeCompile } from '../common/functions/OnBeforeCompilePlugin'
 import { ObjectLayers } from '../scene/constants/ObjectLayers'
-import { EffectMapType, defaultPostProcessingSchema } from '../scene/constants/PostProcessing'
 import { WebXRManager, createWebXRManager } from '../xr/WebXRManager'
 import { XRState } from '../xr/XRState'
 import { RendererState } from './RendererState'
 import WebGL from './THREE.WebGL'
+import { EffectMapType, defaultPostProcessingSchema } from './effects/PostProcessing'
 import { updateShadowMap } from './functions/RenderSettingsFunction'
 import { changeRenderMode } from './functions/changeRenderMode'
 import { configureEffectComposer } from './functions/configureEffectComposer'
