+++ packages/engine/src/renderer/components/InfiniteGridHelper.ts
@@ -50,9 +50,9 @@ import { getMutableState, useHookstate } from '@etherealengine/hyperflux'
 import { useEffect } from 'react'
 import { NameComponent } from '../../common/NameComponent'
 import { GroupComponent, addObjectToGroup, removeObjectFromGroup } from '../../scene/components/GroupComponent'
+import { setObjectLayers } from '../../scene/components/ObjectLayerComponent'
 import { setVisibleComponent } from '../../scene/components/VisibleComponent'
 import { ObjectLayers } from '../../scene/constants/ObjectLayers'
-import { setObjectLayers } from '../../scene/functions/setObjectLayers'
 import { RendererState } from '../RendererState'
 import LogarithmicDepthBufferMaterialChunk from '../constants/LogarithmicDepthBufferMaterialChunk'
