+++ packages/engine/src/renderer/functions/configureEffectComposer.ts
@@ -48,10 +48,10 @@ import { ShaderPass } from 'postprocessing'
 import { EngineState } from '../../EngineState'
 import { CameraComponent } from '../../camera/components/CameraComponent'
 import { ObjectLayers } from '../../scene/constants/ObjectLayers'
-import { EffectMap, EffectPropsSchema, Effects } from '../../scene/constants/PostProcessing'
 import { HighlightState } from '../HighlightState'
 import { RendererState } from '../RendererState'
 import { EffectComposerWithSchema, EngineRenderer, PostProcessingSettingsState } from '../WebGLRendererSystem'
+import { EffectMap, EffectPropsSchema, Effects } from '../effects/PostProcessing'
 import { SDFSettingsState } from '../effects/sdf/SDFSettingsState'
 import { SDFShader } from '../effects/sdf/SDFShader'
 import { changeRenderMode } from './changeRenderMode'
