+++ packages/engine/src/scene/components/HyperspaceTagComponent.ts
@@ -65,8 +65,8 @@ import { ObjectDirection } from '../../common/constants/Axis3D'
 import { createTransitionState } from '../../common/functions/createTransitionState'
 import { TransformComponent } from '../../transform/components/TransformComponent'
 import { SceneLoadingSystem } from '../SceneModule'
+import { setObjectLayers } from '../components/ObjectLayerComponent'
 import { ObjectLayers } from '../constants/ObjectLayers'
-import { setObjectLayers } from '../functions/setObjectLayers'
 import { GroupComponent, addObjectToGroup } from './GroupComponent'
 import { PortalComponent, PortalEffects, PortalState } from './PortalComponent'
 import { VisibleComponent } from './VisibleComponent'
