+++ packages/engine/src/scene/components/MediaComponent.ts
@@ -50,11 +50,11 @@ import { removePannerNode } from '../../audio/PositionalAudioFunctions'
 import { NameComponent } from '../../common/NameComponent'
 import { RendererState } from '../../renderer/RendererState'
 import { EngineRenderer } from '../../renderer/WebGLRendererSystem'
+import { setObjectLayers } from '../components/ObjectLayerComponent'
 import { ObjectLayers } from '../constants/ObjectLayers'
 import { PlayMode } from '../constants/PlayMode'
 import { addError, clearErrors, removeError } from '../functions/ErrorFunctions'
 import isHLS from '../functions/isHLS'
-import { setObjectLayers } from '../functions/setObjectLayers'
 import { addObjectToGroup } from './GroupComponent'
 import { setVisibleComponent } from './VisibleComponent'
