+++ packages/engine/src/scene/components/MountPointComponent.ts
@@ -35,8 +35,8 @@ import { EntityTreeComponent } from '@etherealengine/engine/src/transform/compon
 import { NameComponent } from '../../common/NameComponent'
 import { matches, matchesVector3 } from '../../common/functions/MatchesUtils'
 import { RendererState } from '../../renderer/RendererState'
+import { setObjectLayers } from '../components/ObjectLayerComponent'
 import { ObjectLayers } from '../constants/ObjectLayers'
-import { setObjectLayers } from '../functions/setObjectLayers'
 import { addObjectToGroup } from './GroupComponent'
 import { setVisibleComponent } from './VisibleComponent'
