+++ packages/engine/src/scene/components/ObjectGridSnapComponent.ts
@@ -54,8 +54,8 @@ import {
 } from 'three'
 import { NameComponent } from '../../common/NameComponent'
 import { computeTransformMatrix } from '../../transform/systems/TransformSystem'
+import { setObjectLayers } from '../components/ObjectLayerComponent'
 import { ObjectLayers } from '../constants/ObjectLayers'
-import { setObjectLayers } from '../functions/setObjectLayers'
 import { addObjectToGroup, removeObjectFromGroup } from './GroupComponent'
 import { MeshComponent } from './MeshComponent'
 import { ObjectLayerMaskComponent } from './ObjectLayerComponent'
