+++ packages/engine/src/scene/components/PointLightComponent.ts
@@ -37,8 +37,8 @@ import { mergeBufferGeometries } from '../../common/classes/BufferGeometryUtils'
 import { matches } from '../../common/functions/MatchesUtils'
 import { RendererState } from '../../renderer/RendererState'
 import { isMobileXRHeadset } from '../../xr/XRState'
+import { setObjectLayers } from '../components/ObjectLayerComponent'
 import { ObjectLayers } from '../constants/ObjectLayers'
-import { setObjectLayers } from '../functions/setObjectLayers'
 import { addObjectToGroup, removeObjectFromGroup } from './GroupComponent'
 import { setVisibleComponent } from './VisibleComponent'
