+++ packages/engine/src/scene/components/SpawnPointComponent.ts
@@ -36,8 +36,8 @@ import { AssetLoader } from '../../assets/classes/AssetLoader'
 import { NameComponent } from '../../common/NameComponent'
 import { matches } from '../../common/functions/MatchesUtils'
 import { RendererState } from '../../renderer/RendererState'
+import { setObjectLayers } from '../components/ObjectLayerComponent'
 import { ObjectLayers } from '../constants/ObjectLayers'
-import { setObjectLayers } from '../functions/setObjectLayers'
 import { addObjectToGroup } from './GroupComponent'
 import { setVisibleComponent } from './VisibleComponent'
