+++ packages/engine/src/xrui/functions/createXRUI.tsx
@@ -37,9 +37,9 @@ import { Entity } from '@etherealengine/ecs/src/Entity'
 import { createEntity, EntityContext } from '@etherealengine/ecs/src/EntityFunctions'
 import { InputComponent } from '../../input/components/InputComponent'
 import { addObjectToGroup } from '../../scene/components/GroupComponent'
+import { setObjectLayers } from '../../scene/components/ObjectLayerComponent'
 import { VisibleComponent } from '../../scene/components/VisibleComponent'
 import { ObjectLayers } from '../../scene/constants/ObjectLayers'
-import { setObjectLayers } from '../../scene/functions/setObjectLayers'
 import { DistanceFromCameraComponent } from '../../transform/components/DistanceComponents'
 import { XRUIComponent } from '../components/XRUIComponent'
 import { XRUIStateContext } from '../XRUIStateContext'
